uplevelling data finding to gwftools submodule
